requirements.txt
alembic==1.9.4
anyio==3.6.2
asyncclick==8.1.3.4
asyncpg==0.27.0
bcrypt==4.0.1
boto3==1.26.87
botocore==1.29.87
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
databases==0.7.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
